pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fpdf2>=2.8.3",
    "google-api-python-client>=2.173.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-deepseek>=0.1.3",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.24",
    "langchain-openrouter>=0.0.1",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.8",
    "matplotlib>=3.10.3",
    "openai>=1.88.0",
    "openrouter>=1.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.7",
    "tiktoken>=0.9.0",
    "uv>=0.7.17",
    "xgboost>=3.0.2",
]
